Added emit/perceive semantics
Added emit(message) functionality to Thing class. The message is passed to the container and any other contained objects, if they support the perceive() method

Player now stores an attached console, specified in __init__()

Added perceive(message) method to Player. The message is written to the attached console.

Added emits when a creature enters and leaves a room.

Added dummy perceive(msg) method to Creature.