Greatly simplified the verb "validation" concept.
Previously every verb function had to support a "validate" flag, which if set meant the verb function didn't actually DO anything but just returned "True" if it was valid to call the verb with this direct object etc. This led to a lot of ugly and redundant code in verb functions.

The new concept is much simpler: verb functions return True if they handle the command, and an error message string if they don't.

The parser finds all objects with verb functions matching the verb the user typed, then tries each of them starting with the direct object (if any). If a object's verb function returns True the command is handled, otherwise the parser prints the error message from the FIRST object it tried.

This isn't perfect. For example, suppose there is a rusty sword in the room and a vorpal sword in the user's inventory. The user types "swing sword". The parser finds two swords that support the "swing" verb. Say the programmer has the verb function check whether the sword is in the user's inventory, and print an error message if not: "You must be holding the sword to swing it!".

Ideally, the system would validate the verb "swing" on both swords, see that it is invalid on the one in the room, and remove it from the list of possible objects. This is what the previous version did, but the code complications were not worth it.  Instead the new version will first see that there are two swords matching the verb, and ask "Did you mean the rusty sword or the vorpal sword?".