Fixed "dinner plate" bug, and minor others.
For some reason "Look plate", "examine plate", etc didn't work.

Turned out to be the processing in diagram_sentence that parsed out prepositions.

Was supposed to look for preposition strings with a spoace before and after, e.g. " on " or " under ". But I had added "with" and "at" without the spaces, which meant it was finding the "at" in plate and spliting the sentence there. As a result the string for the direct object was returned as "pl" instead of "plate" causing the odd behavior.

Also changing the "look at" functions to return an error string if a direct object is provided but it doesn't match the supporting verb object. The Room class will support "look" as an intransitive verb, other Things should only support the transitive usage.

Also noticed that the player was included twice in possible_objects via  [user] and user.location.contents