Reworked save/load of entire game.
A major rework of how the game state is pickled and unpickled for saving and restoring saved versions of the game.

The main idea is that when objects refer to other objects, such as in the location and contents fields, we store our unqiue ID strings rather than the reference to the object itself. Then when we load the game we re-build the Thing.ID_dict, which maps from ID strings to objects, with the newly loaded objects.

This is implemented via custom `__getstate__()` and `__setstate__()` functions for the Thing and Player class. In the future we may want to write similar functions for other objects that inherit Thing and add new references to other objects (e.g. the Room class refers to other room objects via the `exits[]` attribute).

This is all preparatory to implementing save and load player functionality, which will need to load a player into a persistent ongoing game.