More progress on saving and loading.
Savegame and loadgame now work again. Under the hood the code is cleaner, and better prepared for save and load player.

One problem was introduced by the Player.start_loc field, which stored a reference to the room where a player will go if they die. Since pickle and unpickle recursively act on any objects referenced by an object, the start location was getting duplicated in a funny way that threw the game off. Avoided this for now simply by storing the ID of the starting room instead.

Another problem was the Room.exits[] list, which also stored refences to the Room objects themselves and again invoked recursive pickling. This problem was solved by storing IDs in he pickling process (using __getstate__()) and converting them back to object references after unpickling (using _restore_objs_from_IDs())

Save and load player still don't work, but are getting closer - the main problem is that when you load a player from a saved file, anything he/she is carrying (for example, the flashlight) is a new object and needs to be added to the game - even if there is already a copy of the same object already in the game. Figuring out how to unpickle objects a player is carrying is the challenge, since these objects will need to be created from scratch and any IDs that were stored (referring for example to contents of a container, or location of things in a container) need to be adjusted to point to the new, local copies of those objects.