Cloning objects almost works but not for Scenery
An object is created from it's Class connstructor, i.e. calling __init__(), and then all the attributes are set. However it doesn't deal with attributes that are references to objects, or to Actions, so you only get the Actions that are defined in __init__(). This means it currently doesn't re-create Actions associated with the object after its creation (which currently only happens with Scenery objects and the add_response() function).