Added test code for client-server interaction.
I found this HOWTO writeup very helpful:

https://docs.python.org/3/howto/sockets.html#socket-howto

It points out that the way sockets are normally used, a "client socket" is created and used exactly once: it sends a request, and possibly receives some data, and then shuts down. Think about a web browser sending http requests and a web server handling them and sending back the data for the web page.

This probably explains why our example is failing after the first connection succeeds - the socketserver library probably implements things in this conventional fashion.

The HOWTO above explains how to use raw sockets (so not the socketserver library) to get around this issue, transferring data back and forth via multiple send() and recv() calls over time. It's a bit ugly, but not terrible. That may be our best solution.

Another option might be to use socketserver and go ahead and have the client open a new socket for every command the user types. THIS IS WHAT I'VE CURRENTLY GOT WORKING IN THE CHECKED-IN CODE. I think the main problem there is, how is the server supposed to know which client is typing? One possibility would be to have the client code prepend every command with the name of the player. That might not be so bad.

One weird thing, if you run the server and quit it "cleanly" (by passing it a "quit" command in my current implementation), then you can't run it again for about 30-60 seconds - you get a "OSError: [Errno 48] Address already in use" error. If you just wait a while, the error goes away. Try running socket_test.py in one terminal window and client.py in another to see what I mean. I can't figure out how to close the socket cleanly enough that this address is released sooner. Might have to do with a default timeout value somewhere, in which case the problem may go away when we implement non-blocking calls.