An interesting example of converting to new action system
This very simple object just creates a new `Thing` object in the
`clone()` function, and adds an Action whose verb function is
also defined in the module. Since there is no class with which to
associate the `actions[]` dictionary, we have to make a copy of
`actions[]` associated with the object rather than with Thing.
Otherwise the new action we add to the dictionary will appear to
all instances of Thing or any subclass of Thing.