test_grid() now takes exit_probability as a paramter, and prints a map.
The major bug was inconsistently treating i and j as x and y. This is a
rookie mistake, very common - when we think of i,j as indexing into an
array we tend to think of i as the row and j as the column, with increasing
rows going down the page. But when we think of x and y as coordinates
on the plane, we tend to think of x as horizontal and y as vertical, with
increasing y going up the page. Once I got this right the maps printed
nicely and illustrated the sparsity of the maze structure. Improving this
will be a fun/tricky problem!