Improved ordinal usage to handle pluralities
Specifying an object using an ordinal ("first", "third", "tenth", etc.)
simply picked the corresponding number from a list of objects. But
when those objects are pluralities, it would get thrown off. This commit
modifies the code to walk through the list of possible matching objects,
keeping track of quantities due to pluralities.