Support string literals in player commands
Players can now type commands with strings, i.e. sections of text
between double-quote marks `"`. The parser will treat those sections
as a single "word" to be passed to the action verb, and will not
convert them to lowercase or strip out articles ('a', 'an', 'the').

In the process, streamlined & simplified these early stages of parsing.