Removed special handling of "go in" from parse()
This code was a special case to allow the user to type "go in" without
triggering the use of "in" as a preposition. But I think we should just
avoid the use of "in" as an exit name, and when necessary overload
the "go" verb in Room to handle phrases like "go in the cave". We
can also use the descriptive text to steer players towards language
like this, or the "enter" verb, or other solutions. @1Bayshore do you
agree with this solution? And can you find/fix any places where we
use "in" as an exit name?