Supporting plurals implemented, almost debugged
You can now type things like, "drop ten gold coins", and if you
are carrying thirty gold, it will split off ten of them to drop.
The remaining known problems are:
- if the action doesn't change the ten coins in any meaningful way
(e.g. looking at coins wouldn't affect them), the ten coins should be
merged back into the other twenty. This isn't implemented yet.
- Conflicts with the code in `_try_verb_from_obj()`, which splits
off a singleton if the object is a plurality,

The solution is probably to move all the splitting- and merging-
pluralities code in `_try_verb_from_obj()` outside the that function
and have it handled by the `find_matching_objects()` code and
a generic cleanup phase that merges any objects that are identical
(including location) after all parsing is done. This sounds difficult
and may be tricky to implement efficiently, but actually shouldn't
be bad.