IMPORTANT: change default __str__ for Thing class
See github issue #133
https://github.com/davedotluebke/old-skool-text-game/issues/133

Basically a common pattern is to count on the __str__() behavior
of Thing objects when constructing messages or descriptions that
players see, and another common pattern is to construct perceive
semantics (see github #8) using the player's id. By changing the
default behavior of `Thing.id()` we hope to make the second pattern
easier, more intuitive, and less bug-prone -- and hope that it won't
break too many places using the first pattern.