Started ACL access work; added "groups" command
We will use access control lists (ACLs) to control wizard and
administrator privileges. ACLs maintain a set of "roles" each of
which has certain "permissions". From the player perspective
these roles are called 'groups' and are controlled with the
`groups` command.

The "groups" command is complex. It is used to list or modify
wizard-group membership. Still in the process of debugging.

- added `wizards` and `roles` dictionarys to Game, and methods
to keep them in sync.
- added `player_roles.json` file to store these dictionaries
- added `groups` command to Player.
- added `checkPlayerExists()` to gametools
- changed `code` in CSS to make code pretty-printing work.