"debug all" now logs all game messages to console
Modified ConHandler to maintain a list of consoles that will receive
log messages handed by the Handler. Created a new ConsHandler
that is attached to every logger created for every object in the game.
The "debug all" command attaches the user's console to that
ConsHandler, effectively subscribing the user to all log messages
created. After the specified number of seconds have elapsed (60 by
default) the user's console is removed from the handler.

Removed the "printing handler" which was intended to log all
messages to stderr, since this new functionality is better.